Changed: Repair command will now do the PreFireChecks once again after the repair timer has expired and before something realy is repaired to ensure that all required conditions are still right.
Changed: House Transfer to Guild, added several messages and an additional check if the transfering player is leader of the guild.

Added: New Serverproperty INDOOR_ITEMS_DEPEND_ON_SIZE, if this is true the max amount of indoor items will depend on the house size and be 40, 60, 80, 100 like on live. If this is false the old MAX_INDOOR_ITEMS value is used.

git-svn-id: https://dolserver.svn.sourceforge.net/svnroot/dolserver/DOLSharp/trunk@1220 6f85df0a-d70d-0410-9e71-934ce63efd45